% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimation}
\alias{estimation}
\title{Cross-fitting estimation of the GATEs}
\usage{
estimation(tra, est, y, covars, type, minsize, cp)
}
\arguments{
\item{tra}{training sample}

\item{est}{estimation sample}

\item{y}{relevant interval of Y}

\item{covars}{provides names of covariables}

\item{type}{indicates whether we consider D=1 or D=0}

\item{minsize}{pruned tree insists on at least 2*"minsize" observations in the additional trees to identify the promising subgroups}

\item{cp}{sets complexity parameter which rpart uses to fit the tree before pruning; default=0}
}
\value{
augmented inverse probability weighting scores, indicator for promising leaves
}
\description{
Grow pruned tree to identify heterogeneous subgroups and picks promising leaves (based on training data)
Calculate magnitude of potential violations within promising leaves (in estimation sample)
}
