% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gates}
\alias{gates}
\title{Calculates the Group Average Treatment Effects for leaves only}
\usage{
gates(sample, treeinuse, ordered_frame)
}
\arguments{
\item{sample}{data to use}

\item{treeinuse}{tree to use}

\item{ordered_frame}{ordered tree frame}
}
\description{
Calculates the Group Average Treatment Effects for leaves only
}
