% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gates_nodes}
\alias{gates_nodes}
\title{Calculates the Group Average Treatment Effects for all nodes (terminal and non-terminal)}
\usage{
gates_nodes(sample, treeinuse, ordered_frame)
}
\arguments{
\item{sample}{data to use}

\item{treeinuse}{tree to use}

\item{ordered_frame}{ordered tree frame}
}
\description{
Calculates the Group Average Treatment Effects for all nodes (terminal and non-terminal)
}
