% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LATEtest.R
\name{LATEtest}
\alias{LATEtest}
\title{LATEtest}
\usage{
LATEtest(data, covars, huge = FALSE, tree_fraction = 0.5,
  minsize = 100, cp = 0, subsets = 4, alpha = 0.05, seed = 10101)
}
\arguments{
\item{data}{provides dataset: outcome must be labelled Y, treatment D and instrument Z}

\item{covars}{provides names of covariables in data (e.g., "Xvar1" "Xvar2" "Xvar3")}

\item{huge}{if set to TRUE, model for orthogonalization learned on random subset of data (size defined in tree_fraction)}

\item{tree_fraction}{fraction of the data used to build each tree of causal forest (and if huge==T, also used for regression forests); default=0.5}

\item{minsize}{causal forest insists on at least "minsize" treated and "minsize" control observations per leaf, and pruned tree insists on at least 2*"minsize" observations in the additional trees to identify the promising subgroups}

\item{cp}{sets complexity parameter which rpart uses to fit the tree before pruning; default=0}

\item{subsets}{number of subsets used to discretize the outcome (using an equidistant grid from the minimum to the maximum value of Y)}

\item{alpha}{nominal significance level of the test}

\item{seed}{set random seed number}
}
\value{
list of the pruned trees, test results
}
\description{
Main function: Prepares data, calls help functions, stores results.
See the LATEtest website on \url{https://github.com/farbmacher/LATEtest} for more information, documentation and examples.
Please report any bugs to \email{farbmacher@econ.lmu.de}
}
\examples{
\dontrun{
library("LATEtest")
library("mvtnorm", "rpart.plot")

# Generate data:
n = 3000; p = 3; rho=0.3
u <- rnorm(n)
v <- rho * u + sqrt(1 - rho^2) * rnorm(n)
X <- matrix(rnorm(n * p), n, p)
colnames(X) <- paste("Xvar", 1:p, sep="")
Z <- rbinom(n, size = 1, prob = 0.5)
D<-as.numeric(0.2 * Z + v > 0)
# local violation of the exclusion restriction:
gamma <- as.numeric(ifelse(X[, 2] < -1, 1.25, 0))
Y <- as.vector(D + gamma * Z + u)
data <- as.data.frame(cbind(Y,D,Z,X))

# Perform test:
covars = paste0(colnames(data)[4:ncol(data)])
test <- LATEtest(data = data, covars = covars, subsets = 4, alpha = 0.05)
test

# Draw plot of pruned tree that led to local violation of LATE assumptions:
maxtree <- eval(parse(text = paste("test$treelist$tree_", test$maxTtree$label, test$maxTtree$J,sep = "")))
rpart.plot::rpart.plot(maxtree, extra = 101, box.palette = "GyRd", shadow.col = "gray", nn = TRUE, roundint = FALSE)
}
}
