% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oob}
\alias{oob}
\title{Constructs pseudo-outcome, runs regression and causal forests and predicts out-of-bag (oob)}
\usage{
oob(est, y, huge, tree_fraction, minsize, type)
}
\arguments{
\item{est}{estimation sample}

\item{y}{relevant interval of Y}

\item{huge}{if set to TRUE, model for orthogonalization learned on random subset of data (size defined in tree_fraction)}

\item{tree_fraction}{fraction of the data used to build each tree of causal forest (and if huge==T, also used for regression forests); default=0.5}

\item{minsize}{causal forest insists on at least "minsize" treated and "minsize" control observations per leaf}

\item{type}{indicates whether we consider D=1 or D=0}
}
\value{
pseudo-outcome, oob predictions of pseudo-outcome, oob estimates of causal forest
}
\description{
Constructs pseudo-outcome, runs regression and causal forests and predicts out-of-bag (oob)
}
