% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row2leaf}
\alias{row2leaf}
\title{Converts row number from tree$frame into leaf number}
\usage{
row2leaf(row, rp)
}
\arguments{
\item{row}{row}

\item{rp}{rp}
}
\description{
Converts row number from tree$frame into leaf number
}
